browser-use-milvus/requirements.txt
browser-use>=0.0.5
pymilvus>=2.3.6
langchain>=0.1.0
langchain-google-genai>=0.0.6
python-dotenv>=1.0.0
fastapi>=0.109.2
uvicorn>=0.27.1
pydantic>=2.6.1
python-multipart>=0.0.9
sentence-transformers>=2.5.1 